scale/scale/modeler/requirements.txt
aiohttp
grpcio-tools==1.62.3
grpcio==1.62.3
numpy
opentelemetry-proto==1.27.0
pandas
protobuf==4.25.5
pyyaml
scikit_learn
scikit_learn_extra
scipy
tabulate
tenacity
watchdog